Provide an API appropriate for observing model updates in actors
The problem:
The Queried macro is not useful in actors because actors must own their own ModelContext, but since that model update is private it'll never have external updates.

Solution:
Observe the main context for updates, then perform fetches in the private model context to perform any work related to said updates in the background.